+++ Source/AIM/Common/OrientationOps/HexagonalOps.cpp
@@ -33,14 +33,14 @@
 #include "AIM/Common/AIMMath.h"
 
 
-  const static double m_pi = M_PI;
-  static const double HexDim1InitValue = pow((0.75*((m_pi/2.0)-sin((m_pi/2.0)))),(1.0/3.0));
-  static const double HexDim2InitValue = pow((0.75*((m_pi/2.0)-sin((m_pi/2.0)))),(1.0/3.0));
-  static const double HexDim3InitValue = pow((0.75*((m_pi/6.0)-sin((m_pi/6.0)))),(1.0/3.0));
-  static const double HexDim1StepValue = HexDim1InitValue/18.0;
-  static const double HexDim2StepValue = HexDim1InitValue/18.0;
-  static const double HexDim3StepValue = HexDim1InitValue/6.0;
-static const double HexQuatSym[12][5] = {{0.000000000, 0.000000000, 0.000000000, 0.000000000, 1.000000000},
+  const static float m_pi = M_PI;
+  static const float HexDim1InitValue = pow((0.75*((m_pi/2.0)-sin((m_pi/2.0)))),(1.0/3.0));
+  static const float HexDim2InitValue = pow((0.75*((m_pi/2.0)-sin((m_pi/2.0)))),(1.0/3.0));
+  static const float HexDim3InitValue = pow((0.75*((m_pi/6.0)-sin((m_pi/6.0)))),(1.0/3.0));
+  static const float HexDim1StepValue = HexDim1InitValue/18.0;
+  static const float HexDim2StepValue = HexDim1InitValue/18.0;
+  static const float HexDim3StepValue = HexDim1InitValue/6.0;
+static const float HexQuatSym[12][5] = {{0.000000000, 0.000000000, 0.000000000, 0.000000000, 1.000000000},
                        {0.000000000, 0.000000000, 0.000000000, 0.500000000, 0.866025400},
                        {0.000000000, 0.000000000, 0.000000000, 0.866025400, 0.500000000},
                        {0.000000000, 0.000000000, 0.000000000, 1.000000000, 0.000000000},
@@ -52,7 +52,7 @@ static const double HexQuatSym[12][5] = {{0.000000000, 0.000000000, 0.000000000,
                        {0.000000000, 0.000000000, 1.000000000, 0.000000000, 0.000000000},
                        {0.000000000, -0.50000000, 0.866025400, 0.000000000, 0.000000000},
                        {0.000000000, -0.86602540, 0.500000000, 0.000000000, 0.000000000}};
-static const double HexRodSym[12][3] = {{0,0,0},
+static const float HexRodSym[12][3] = {{0,0,0},
                     {0, 0, 0.57735},
                     {0, 0, 1.73205},
                     {0, 0, 10000000000.0},
@@ -74,10 +74,10 @@ HexagonalOps::~HexagonalOps()
 {
 }
 
-double HexagonalOps::getMisoQuat( double q1[5],double q2[5],double &n1,double &n2,double &n3)
+float HexagonalOps::getMisoQuat( float q1[5],float q2[5],float &n1,float &n2,float &n3)
 {
   int numsym = 0;
-  double quatsym[24][5];
+  float quatsym[24][5];
 
   numsym = 12;
   for (int i = 0; i < 12; i++)
@@ -92,9 +92,9 @@ double HexagonalOps::getMisoQuat( double q1[5],double q2[5],double &n1,double &n
 }
 
 
-void HexagonalOps::getFZRod(double &r1,double &r2, double &r3)
+void HexagonalOps::getFZRod(float &r1,float &r2, float &r3)
 {
-  double rodsym[24][3];
+  float rodsym[24][3];
   int numsym;
   numsym = 12;
   for(int i=0;i<12;i++)
@@ -107,10 +107,10 @@ void HexagonalOps::getFZRod(double &r1,double &r2, double &r3)
   _calcFZRod(rodsym, numsym,r1, r2, r3);
 }
 
-void HexagonalOps::getNearestQuat( double *q1, double *q2)
+void HexagonalOps::getNearestQuat( float *q1, float *q2)
 {
   int numsym = 0;
-  double quatsym[24][5];
+  float quatsym[24][5];
 
   numsym = 12;
      for(int i=0;i<12;i++)
@@ -124,10 +124,10 @@ void HexagonalOps::getNearestQuat( double *q1, double *q2)
     _calcNearestQuat(quatsym, numsym, q1, q2);
 }
 
-void HexagonalOps::getFZQuat(double *qr)
+void HexagonalOps::getFZQuat(float *qr)
 {
   int numsym = 0;
-  double quatsym[24][5];
+  float quatsym[24][5];
 
   numsym = 12;
   for(int i=0;i<12;i++)
@@ -142,10 +142,10 @@ void HexagonalOps::getFZQuat(double *qr)
 }
 
 
-int HexagonalOps::getMisoBin(double n1, double n2, double n3)
+int HexagonalOps::getMisoBin(float n1, float n2, float n3)
 {
-  double dim[3];
-  double bins[3];
+  float dim[3];
+  float bins[3];
 
   dim[0] = HexDim1InitValue;
   dim[1] = HexDim2InitValue;
@@ -158,11 +158,11 @@ int HexagonalOps::getMisoBin(double n1, double n2, double n3)
 }
 
 
-void HexagonalOps::determineEulerAngles(int choose, double &synea1, double &synea2, double &synea3)
+void HexagonalOps::determineEulerAngles(int choose, float &synea1, float &synea2, float &synea3)
 {
-  double init[3];
-  double step[3];
-  double phi[3];
+  float init[3];
+  float step[3];
+  float phi[3];
 
   init[0] = HexDim1InitValue;
   init[1] = HexDim2InitValue;
@@ -178,11 +178,11 @@ void HexagonalOps::determineEulerAngles(int choose, double &synea1, double &syne
 }
 
 
-void HexagonalOps::determineHomochoricValues( int choose, double &r1, double &r2, double &r3)
+void HexagonalOps::determineHomochoricValues( int choose, float &r1, float &r2, float &r3)
 {
-  //double init1, init2, init3;
-  double step[3];
-  double phi[3];
+  //float init1, init2, init3;
+  float step[3];
+  float phi[3];
 
   step[0] = HexDim1StepValue/2.0;
   step[1] = HexDim2StepValue/2.0;
@@ -194,10 +194,10 @@ void HexagonalOps::determineHomochoricValues( int choose, double &r1, double &r2
   return _calcDetermineHomochoricValues(step, phi, choose, r1, r2, r3);
 }
 
-int HexagonalOps::getOdfBin(double r1, double r2, double r3)
+int HexagonalOps::getOdfBin(float r1, float r2, float r3)
 {
-  double dim[3];
-  double bins[3];
+  float dim[3];
+  float bins[3];
 
   dim[0] = HexDim1InitValue;
   dim[1] = HexDim2InitValue;
